% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\docType{data}
\name{gdpPercap}
\alias{gdpPercap}
\title{Random Forest Cross-Validation}
\format{An object of class \code{NULL} of length 0.}
\usage{
gdpPercap
}
\arguments{
\item{k}{Number of folds.}
}
\value{
A numeric with the cross-validation error.
}
\description{
Predicts outputs Sepal.Length using covariates
Sepal.Width, Petal.Length, and Petal.Width by random forest methods.
}
\examples{
my_rf_cv(5)

}
\keyword{cross-validation}
