% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{linear models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A class object}

\item{data}{A set of data}
}
\value{
A dataframe of summary with rows for each coefficient and columns for the
Estimate, Std. Error, t value, and Pr(>|t|).
}
\description{
my_lm is used to fit linear models.
}
\examples{
my_lm(mpg ~ hp + wt, mtcars)

}
\keyword{prediction}
