% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pow.R
\name{my_pow}
\alias{my_pow}
\title{Power function}
\usage{
my_pow(x, power = 2)
}
\arguments{
\item{x}{Numeric input to be raised to the power of \code{power}.}

\item{power}{Numeric input for the power that \code{x} will be raised to,
defaults to \code{2}.}
}
\value{
Numeric representing \code{x} raised to the power of \code{power}.
}
\description{
This function raises input to a power.
}
\examples{
my_pow(4)
my_pow(4, power = 3)

}
