% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Integer representing Number of folds.}
}
\value{
A numeric with the cross-validation error.
}
\description{
Predicts penguin species using random forest methods.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
