% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{KNN Cross-Validatin}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{A data frame that contains the true class value of the
input training data \code{train}.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list containing the following objects:
\itemize{
\item class: objects class that contains a vector of the predicted
class for all observations.
\item cv_error: a numeric represents cross-validation misclassification error.
}
}
\description{
It uses the k-nearest-neighbors method to predict the class.
}
\examples{

my_knn_cv(penguins[, -5], penguins[, 5], 5, 1)

}
\keyword{prediction}
